build(pom): 升级公共包

+++ team4u-kv/src/main/java/org/team4u/kv/KeyValueRepository.java
@@ -57,7 +57,7 @@ public interface KeyValueRepository {
     /**
      * 统计kv数量
      */
-    int count(String type);
+    long count(String type);
 
     /**
      * 检查是否包含指定标识数据
